+++ test/devote/blockchain/blockchainconfigwithnodefaultctr/blockchainwithmissingimplementation/BlockchainWihMissingImplementationIssuerOperation.java
@@ -16,7 +16,7 @@ public class BlockchainWihMissingImplementationIssuerOperation implements Issuer
     }
 
     @Override
-    public int calcNumOfAccountsNeeded(long totalVotesCap) {
-        return 0;
+    public long calcNumOfAccountsNeeded(long totalVotesCap) {
+        return 0L;
     }
 }
